Backport 'fix integer overflow in monitor refresh rate calculation' to GTK3
Original MR by Volker Rümelin.

See https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/2884

+++ gdk/win32/gdkmonitor-win32.c
@@ -523,7 +523,7 @@ populate_monitor_devices_from_display_config (GPtrArray *monitors)
 
       refresh = &dispconf_paths[path_index].targetInfo.refreshRate;
       gdk_monitor_set_refresh_rate (mon,
-                                    refresh->Numerator * 1000 / refresh->Denominator);
+                                    refresh->Numerator * (UINT64) 1000 / refresh->Denominator);
     }
 
   g_free (dispconf_paths);
